# docs/api-specification.yaml
openapi: 3.0.0
info:
  description: API specification
  version: 0.0.1
  title: Message API specification
servers:
  - url: http://localhost:3000/
paths:
...
  /invitations:
    post:
      description: Invitation creation
      tags:
        - Invitations
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                senderId:
                  type: string
                  required: true
                recipientId:
                  type: string
                  required: true
                content:
                  type: string
                  required: true
      responses:
        201:
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  senderId:
                    type: string
                  recipientId:
                    type: string
                  createdAt:
                    type: string
                  content:
                    type: string
...
  /groups:
    post:
      description: Group creation
      tags:
        - Groups
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  required: true
      responses:
        201:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
...
  /groups/{groupId}/members:
    post:
      description: Adding member to the group
      tags:
        - GroupMembers
      parameters:
        - name: groupId
          in: path
          description: Group ID
          required: true
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                  required: true
      responses:
        201:
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupMember'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Group:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        cover:
          type: string
        groupMembers:
          type: array
    GroupMember:
      type: object
      properties:
        id:
          type: string
        group_id:
          type: string
        user_id:
          type: string
        role:
          type: string